// server/src/types/ErrorCodes.ts
export enum ErrorCodes {
  E001 = 'E001',
  E002 = 'E002',
  E003 = 'E003',
  E004 = 'E004',
  E005 = 'E005',
  E006 = 'E006',
  E007 = 'E007',
  E008 = 'E008',
  E009 = 'E009',
  E010 = 'E010',
  E011 = 'E011',
  E012 = 'E012',
  E013 = 'E013',
  E014 = 'E014',
  E015 = 'E015',
  E016 = 'E016',
  E017 = 'E017',
  E018 = 'E018',
  E019 = 'E019',
  E020 = 'E020',
  E021 = 'E021',
  E022 = 'E022',
  E023 = 'E023',
  E024 = 'E024',
  E025 = 'E025',
  E026 = 'E026',
  E027 = 'E027',
  E028 = 'E028',

  W001 = 'W001',
  W002 = 'W002',
}